Problem 2: Remember the uints.

+++ problem2/problem2.go
@@ -8,21 +8,21 @@ import (
 var PHI = (1 + SQRT5) / 2
 var SQRT5 = math.Sqrt(5)
 
-func fibonacciInverse(f int) int {
+func fibonacciInverse(f uint) uint {
 	if f < 2 {
-		return f
+		return uint(f)
 	}
-	return int(math.Round(math.Log(float64(f)*SQRT5)/math.Log(PHI))) - 1
+	return uint(math.Round(math.Log(float64(f)*SQRT5)/math.Log(PHI))) - 1
 }
 
-func fibonacci(n int) int {
+func fibonacci(n uint) uint {
 	n++
-	return int(math.Round(float64(((math.Pow(PHI, float64(n)) - math.Pow(1-PHI, float64(n))) / SQRT5))))
+	return uint(math.Round(float64(((math.Pow(PHI, float64(n)) - math.Pow(1-PHI, float64(n))) / SQRT5))))
 }
 
-func sum() (sum int) {
+func sum() (sum uint) {
 	max := fibonacciInverse(4000000)
-	for i := 1; i <= max; i++ {
+	for i := uint(1); i <= max; i++ {
 		if (i-2)%3 == 0 {
 			continue
 		}
